feat(): add bike detail with param on vm

+++ app/src/main/java/com/example/motorcyclemanager/data/models/BikeEntity.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class BikeEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val time: Int
+    val time: Float
 )
