+++ app/src/main/java/com/example/motorcyclemanager/presentation/addbike/models/BikeAdded.kt
@@ -1,3 +1,3 @@
 package com.example.motorcyclemanager.presentation.addbike.models
 
-data class BikeAdded (val name: String, val time: Int)
+data class BikeAdded (val name: String, val time: Float)
